requirements.txt
asgiref==3.6.0
boto3==1.26.83
botocore==1.29.83
certifi==2022.12.7
click==8.1.3
colorama==0.4.6
distlib==0.3.6
Django==4.1.7
filelock==3.9.0
Flask==2.2.3
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.2
Pillow==9.4.0
pipenv==2023.2.18
platformdirs==3.0.0
psycopg2==2.9.5
python-dateutil==2.8.2
python-dotenv==1.0.0
s3transfer==0.6.0
six==1.16.0
SQLAlchemy==2.0.4
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2022.7
urllib3==1.26.14
virtualenv==20.19.0
virtualenv-clone==0.5.7
Werkzeug==2.2.3
